IMPORTANT!! READ DESCRIPTION!!
The important change is that I changed the LINK_ENTITY_TO_CLASS functions to the way they're like in CSGO.

As for the rest of the misc changes:
Implemented CBaseHLCombatWeapon since Portal 2 has it and we need a few CVars from the files.
Updated base portal weapons.
Updated portal simulation