Work around VPK crash
You need to call the patched AddVPKFile as early as possible so the VPK list gets flushed and re-filled. If any code tries to access VPK archives >128 before this, it apparently tries to index the file handle array out of bounds and this usually leads to crashes. After flushing the VPK list, this no longer happens.

This could probably be fixed more properly, but I don't currently have the skills/knowledge to be able to figure that out.